Fxi the name bug to rename to original names
