+++ src/components/svg/index.tsx
@@ -1,4 +1,4 @@
-import Menu from './sjkfsjdnfk';
-import QuestionMark from './sklfndslfnds';
+import Menu from './menu';
+import QuestionMark from './questionMark';
 
 export { Menu, QuestionMark };
